# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreateBookingRequest:
    properties:
      contact_email:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      details_json:
        type: string
      end_time:
        type: string
      participants:
        minimum: 1
        type: integer
      place_id:
        type: integer
      service_id:
        type: integer
      start_time:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/sqlc.BookingStatus'
        enum:
        - created
        - confirmed
        - canceled
        - completed
      zone_id:
        type: integer
    required:
    - end_time
    - participants
    - service_id
    - start_time
    - zone_id
    type: object
...
  domain.PatchBookingRequest:
    properties:
      end_time:
        type: string
      participants:
        minimum: 1
        type: integer
      service_id:
        type: integer
      start_time:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/sqlc.BookingStatus'
        enum:
        - created
        - confirmed
        - canceled
        - completed
      zone_id:
        type: integer
    type: object
...
  response.BookingDoc:
    properties:
      created_at:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
      end_time:
        example: "2026-03-20T11:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      participants:
        example: 4
        type: integer
      service_id:
        example: 1
        type: integer
      start_time:
        example: "2026-03-20T10:00:00Z"
        type: string
      status:
        example: created
        type: string
      total_price:
        example: "1500.00"
        type: string
      updated_at:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      zone_id:
        example: 1
        type: integer
    type: object
  response.BookingListResponse:
    properties:
      bookings:
        items:
          $ref: '#/definitions/response.BookingDoc'
        type: array
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
  response.BookingResponse:
    properties:
      booking:
        $ref: '#/definitions/response.BookingDoc'
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
  response.DeleteBookingResponse:
    properties:
      id:
        example: 1
        type: integer
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        example: auth_failed
        type: string
      details: {}
      message:
        example: Authentication failed
        type: string
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
...
  sqlc.BookingStatus:
    enum:
    - created
    - confirmed
    - canceled
    - completed
    type: string
    x-enum-varnames:
    - BookingStatusCreated
    - BookingStatusConfirmed
    - BookingStatusCanceled
    - BookingStatusCompleted
...
host: localhost
info:
  contact: {}
  description: Backend API for PlayGround
  title: PlayGround API
  version: "1.0"
paths:
  /api/v1/auth/{provider}:
    get:
      description: Redirects to provider authorization page
      parameters:
      - description: OAuth provider
        enum:
        - google
        in: path
        name: provider
        required: true
        type: string
      responses:
        "307":
          description: Redirect to provider
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Start OAuth
      tags:
      - auth
  /api/v1/auth/{provider}/callback:
    get:
      description: Completes OAuth flow and redirects with session cookie
      parameters:
      - description: OAuth provider
        enum:
        - google
        in: path
        name: provider
        required: true
        type: string
      responses:
        "302":
          description: Redirect to frontend
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: OAuth callback
      tags:
      - auth
...
  /api/v1/booking:
    get:
      description: Returns all bookings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BookingListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List bookings
      tags:
      - bookings
    post:
      consumes:
      - application/json
      description: Creates a booking from JSON payload (user derived from session)
      parameters:
      - description: Booking payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create booking
      tags:
      - bookings
  /api/v1/booking/{id}:
    delete:
      description: Deletes booking by id
      parameters:
      - description: Booking ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DeleteBookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete booking
      tags:
      - bookings
    get:
      description: Returns booking by id from path param
      parameters:
      - description: Booking ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get booking by id
      tags:
      - bookings
    patch:
      consumes:
      - application/json
      description: Partially updates booking fields by id
      parameters:
      - description: Booking ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Booking patch payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.PatchBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Patch booking
      tags:
      - bookings
...
  /api/v1/user:
    get:
      description: Returns all active users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List users
      tags:
      - users
...
swagger: "2.0"